adapters-ingestion-py/astrocyte-ingestion-document/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "astrocyte-ingestion-document"
version = "0.1.0"
description = "Local document folder poll IngestSource adapter for Astrocyte"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
dependencies = [
    "astrocyte>=0.7.0,<2",
    "pypdf>=4.0",
    "python-docx>=1.1",
    "beautifulsoup4>=4.12",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[project.entry-points."astrocyte.ingest_poll_drivers"]
document = "astrocyte_ingestion_document.source:DocumentFolderIngestSource"

[tool.hatch.build.targets.wheel]
packages = ["astrocyte_ingestion_document"]

[tool.uv.sources]
astrocyte = { path = "../../astrocyte-py", editable = true }

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I"]
